# optimise scanning code