This code manually iterates over `walkdir::WalkDir` from the [walkdir](https://docs.rs/walkdir/) crate to intelligently skip directories and files under certain conditions using [`skip_current_dir`](https://docs.rs/walkdir/2.3.1/walkdir/struct.IntoIter.html#method.skip_current_dir).


**Criterion results**

Before

```
scanner_bench           time:   [5.5149 s 5.6332 s 5.7884 s]
```

After

```
scanner_bench           time:   [135.92 ms 136.49 ms 136.96 ms]
                        change: [-97.578% -97.535% -97.499%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Results in a 97.5% reduction in `scan` runtime on my pc.